Moved straincollection to own file
